Rename car/cdr to head/tail
Since these are just field names which are usually manually set when
doing pattern matching this doesn't really matter (though technically it
is a breaking change).

This is done for consistency with the rest of the file, where head/tail
is used when destructuring. Additionally, this will probably be easier
for students to understand since it is the terminology used for linked
lists in most programming (other than lisp/scheme).