Simplify add_cons_eq_cons_add proof
Only one simp statement is needed. The reason we need this instead of
just marking `has_add.add` and `tuple.add` as `[simp]` is that otherwise
under normal circumstances `simp` is too eager: it expands `has_add.add`
for real numbers as well as for tuples, and it seems like most/all
proofs about addition on real numbers are on `has_add.add` rather than
normal addition. This makes it pretty much impossible to progress.