Make tuples generic on the contained type
Needed for matrices since they'll be tuples of tuples.

Also adds the R^^n notation to indicate an n-tuple of R for more easily
specifying the type of tuples.

We have yet to implement the complex dot product and complex norm, but
we'll get to that when it is needed.