Fix proof that tuples are abelian groups
Now proves that tuples of abelian groups are abelian groups (i.e., G^^n
forms an abelian group when G forms an abelian group). While we could
generalize this to non-abelian groups as well, that is not necessary
here. It shouldn't be hard to do so, however.

It is of note that nsmul and zsmul had to be redefined to the definition
using repeated addition since we can't necessarily multiply by a natural
scalar (even if upcast) on an arbitrary tuple. Although it might have
been more in-line with previous definitions to instead map the
nsmul/zsmul over the elements of the tuple, since this is just an
implementation detail and won't actually be used that's not necessary.

Actually maybe I'll do it later if I'm bored.