Reintroduce inner product spaces
Also defines orthogonality over arbitrary inner product spaces. Note
that the definition of orthogonality does not use the "<<., .>>"
notation for inner products since has_inner.inner defines that notation
separately for real and complex numbers (<<., .>>_R and <<., .>>_C), and
we want orthogonality to be general. We may want to introduce the
general notation in the future.