Create the destruct_tuple tactic
Effectively, for v : a ^ n for n <= m, then running
> destruct_tuple m v using w,
will perform cases on v a total of m times, naming each element as w_x
where x is the appropriate index and _x is the subscript form of x. The
final element, if any, will be called w. Omitting the "using w" will
default to using the current name as the prefix.

Proper documentation is still necessary, but this is in a working state
now. Ideally it would be possible to make it automatically infer the
value of m as being n, but I cannot figure out how to do that currently.